requirement.txt
slack-sdk==3.26.1
python-dotenv==1.0.0
fastapi>=0.100.0
uvicorn>=0.20.0
